Move project configuration pieces into ProjectConfiguration
I originally was opposed to this because it exposes
some of the private workings of the base class to the inheritance chain,
but I decided it was worth the trade off for a cleaner interface in the base.
There is also the possibility that some other use will arise for being able
to define ones own implementation. Possibly a json to INI extension?