Ability to freeze a project or set of projects
During maintenances, etc. we often want to block any new changes
from being pushed into the upstream origin. In some cases this
will apply to just a single repository, but in others all
repositories.

Since this needs to be controlled outside of a given project's
purview, it has to be a system configuration, not a `ProjectConfig`.
...hence a new `Configuration` class